Removes a redundant healthcheck
This healthcheck is cause lots of calls to the email service and causing
them unwanted load.

It also doesn't do what it says it does, there is no actuall checking
that the ID's map correctly and its just checking that the email service
returns anything less than a 500 status. In this case the email service
is returning a 404 because the user doesn't exist and therefor is no way
it could check the ids map correctly.